% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCellTables.R
\name{buildCellTables}
\alias{buildCellTables}
\title{Build all tables related to cells for ToxicoDB and write to disk}
\usage{
buildCellTables(path = "procdata", outDir = "latest", ...)
}
\arguments{
\item{path}{\code{character} path to the directory where the ToxicoSet data was
extracted}

\item{outDir}{\code{character} path to save table csvs to}

\item{...}{\code{pairlist} force further arguments to be named}
}
\description{
Build all tables related to cells for ToxicoDB and write to disk
}
