% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCompoundTables.R
\name{buildCompoundTables}
\alias{buildCompoundTables}
\title{Build all tables related to compound for ToxicoDB and write to disk}
\usage{
buildCompoundTables(
  path = "procdata",
  annotPath = "metadata/drug_annotations.csv",
  outDir = "latest",
  ...,
  updatedCompounds = "metadata/old_newDrugmapping.csv",
  annotColMap = c(compound_id = "compound_id", pubchem = "pubchem", ctd = "ctd", chembl
    = "chembl", drugbank = "drugbank", targets = "targets", carcinogenicity =
    "carcinogenicity", class_in_vivo = "classif_in_vivo", class_in_vitro =
    "classif_in_vitro", class_name = "class_name", smiles = "smiles", inchikey =
    "inchikey", name = "name", NTP = "ntp", IARC = "iarc", DILI_status = "dili_status")
)
}
\arguments{
\item{path}{\code{character} path to the directory where the ToxicoSet data was
extracted}

\item{outDir}{\code{character} path to save table csvs to}

\item{...}{\code{pairlist} force further arguments to be named}

\item{annotColMap}{A \code{vector} named vector mapping the column names in
the annotation .csv to the correct database table names. If we don't
have data for some columns set them to NA and they will be added to the
final .csv as NA columns.}
}
\description{
Build all tables related to compound for ToxicoDB and write to disk
}
