% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMolecProfTables.R
\name{buildMolecProfTables}
\alias{buildMolecProfTables}
\title{Build all tables related to molecular profiles for ToxicoDB and write to disk}
\usage{
buildMolecProfTables(
  path = "procdata",
  outDir = "latest",
  ...,
  molecProf = "molecProf",
  gene = "gene",
  analysis = "analysis"
)
}
\arguments{
\item{path}{\code{character} path to the directory where the ToxicoSet data was
extracted.}

\item{outDir}{\code{character} path to save table csvs to.}

\item{...}{\code{pairlist} force subsequent arguments to be named.}

\item{molecProf}{\code{character} Name of molecular profiles directory within
the \code{path} directory. Default is parameter name (i.e., 'molecProf').}

\item{gene}{\code{character} Name of gene directory within the \code{path} directory.
Default is parameter name.}

\item{analysis}{\code{character} Name of gene directory within the \code{path}
directory. Default is parameter name.}
}
\value{
None. Writes to disk.
}
\description{
\strong{WARNING}: This function uses a lot of RAM. It is not recommended to run
this on systems with less the 32 GB of memory, otherwise it may crash
your system.
}
