% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSensitivityTables.R
\name{buildSensitivityTables}
\alias{buildSensitivityTables}
\title{Build all tables related to samples for ToxicoDB and write to disk}
\usage{
buildSensitivityTables(path = "procdata", outDir = "latest", ...)
}
\arguments{
\item{path}{\code{character} path to the directory where the ToxicoSet data was
extracted}

\item{outDir}{\code{character} path to save table csvs to}

\item{...}{\code{pairlist} force further arguments to be named}
}
\description{
Build all tables related to samples for ToxicoDB and write to disk
}
