% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAllTSets.R
\name{downloadAllTSets}
\alias{downloadAllTSets}
\title{Download all ToxicoSet objects using the ToxicoGx R package.}
\usage{
downloadAllTSets(saveDir = "tSets")
}
\arguments{
\item{saveDir}{\code{character} Path to save the tSets to. Defaults to 'tSets'.}
}
\description{
Download all ToxicoSet objects using the ToxicoGx R package.
}
