% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAnalysisTable.R
\name{extractAllAnalysisTables}
\alias{extractAllAnalysisTables}
\title{Contruct the analysis table for each \code{ToxicoSet} from a list of \code{ToxicoSet}s}
\usage{
extractAllAnalysisTables(tSets, outDir = tempdir())
}
\arguments{
\item{tSets}{\code{list} of \code{ToxicoSet} objects}

\item{outDir}{\code{character} path to save the table to. Defaults to \code{tempdir()}.}
}
\value{
Nothing; writes to disk
}
\description{
Contruct the analysis table for each \code{ToxicoSet} from a list of \code{ToxicoSet}s
}
