% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSensitivityProfileTable.R
\name{extractAllSensitivityProfileTables}
\alias{extractAllSensitivityProfileTables}
\title{Contruct the sensitivityProfile for for each \code{ToxicoSet} from a list of \code{ToxicoSet}s}
\usage{
extractAllSensitivityProfileTables(tSets, outDir = tempdir())
}
\arguments{
\item{tSets}{\code{list} of \code{ToxicoSet} objects}

\item{outDir}{\code{character} path to save the table to. Defaults to \code{tempdir()}.}
}
\value{
Nothing; writes to disk
}
\description{
Contruct the sensitivityProfile for for each \code{ToxicoSet} from a list of \code{ToxicoSet}s
}
