% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSensitivityTable.R
\name{extractAllSensitivityTables}
\alias{extractAllSensitivityTables}
\title{Contruct the sensitivityProfile for for each \code{ToxicoSet} from a list of
\code{ToxicoSet}s}
\usage{
extractAllSensitivityTables(tSets, outDir = tempdir())
}
\arguments{
\item{tSets}{\code{list} of \code{ToxicoSet} objects}

\item{outDir}{\code{character} path to save the table to. Defaults to \code{tempdir()}.}
}
\value{
Nothing; writes to disk
}
\description{
Contruct the sensitivityProfile for for each \code{ToxicoSet} from a list of
\code{ToxicoSet}s
}
