% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAllTables.R
\name{extractAllTables}
\alias{extractAllTables}
\title{Extract all data needed to build the ToxicoDB database tables from a set of
\code{ToxicoSet} objects}
\usage{
extractAllTables(
  path = "tSets",
  pattern = "EM.*rds|.*ldh.*rds|.*drugMatrix.*rds",
  outDir = "procdata"
)
}
\arguments{
\item{path}{\code{character}}
}
\description{
Extract all data needed to build the ToxicoDB database tables from a set of
\code{ToxicoSet} objects
}
