% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSensitivityTable.R
\name{extractSensitivityTable}
\alias{extractSensitivityTable}
\title{Construct the molecularProfile table from a \code{ToxicoSet} object}
\usage{
extractSensitivityTable(tSet, outDir = tempdir(), fileName = name(tSet))
}
\arguments{
\item{tSet}{\code{ToxicoSet} object}

\item{outDir}{\code{character} path to save the table to. Defaults to \code{tempdir()}.}

\item{fileName}{\code{character} name of the output file. Defaults to \code{name(tSet)}.}
}
\value{
Nothing; writes to disk
}
\description{
Construct the molecularProfile table from a \code{ToxicoSet} object
}
