% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCTDpathways.R
\name{getCTDpathways}
\alias{getCTDpathways}
\title{Download pathway data from Comparative Toxicology Database}
\usage{
getCTDpathways(url = "http://ctdbase.org/reports/CTD_genes_pathways.csv.gz")
}
\arguments{
\item{url}{\code{character} Url of the compressed .csv file}
}
\description{
Available data can be found here http://ctdbase.org/downloads/
}
