% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigDBpathways.R
\name{getMSigDBPathways}
\alias{getMSigDBPathways}
\title{Get pathway annotations from MSigDB}
\usage{
getMSigDBPathways(species = c("Homo sapiens", "Rattus norvegicus"), all = TRUE)
}
\arguments{
\item{species}{\code{character} Which species to retrieve the pathway annotations
for.}

\item{all}{\code{logical} Should the data for all species be retrieved?}
}
\value{
A \code{data.table} mapping gene symbol to pathway ids
}
\description{
Get pathway annotations from MSigDB
}
